pyproject.toml
[project]
name = "sudoku-recog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "numpy",
    "kagglehub>=0.4.2",
    "pytest>=9.0.2",
    "torch",
    "markupsafe==2.1.5",
    "pillow>=12.0.0",
    "matplotlib>=3.10.8",
    "opencv-python>=4.13.0.90",
    "imutils>=0.5.4",
]

[project.optional-dependencies]
docs = [
    "ipykernel",
    "ipython",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu130"}
]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
